requirements.txt
streamlit
transformers
sentencepiece
python-dotenv
torch
